requirements.txt
#streamlit req
setuptools>=65.0.0
wheel>=0.38.0

# Core ML requirements
numpy==1.24.3
scipy>=1.10.1
torch>=2.1.0
coremltools>=6.3.0

# Visualization requirements
matplotlib>=3.7.1
seaborn>=0.12.2
plotly>=5.0.0
psutil>=5.9.0  # For system monitoring and activity detection

# Training utilities
tqdm>=4.65.0  # For progress bars
tensorboard>=2.14.0  # For training monitoring
pandas>=2.0.3  # For data handling
pyarrow>=12.0.0  # For parquet file storage
fastparquet>=0.8.0  # Alternative parquet backend

# Compression libraries
lz4>=4.3.0  # For tiered storage compression

# CLI interface
click>=8.0.0  # For command-line interface
PyYAML>=6.0.0  # YAML config

# Testing
pytest>=7.4.0  # For running tests

# Dashboard web framework
Flask>=2.3.0  # Web framework for metrics dashboard
Flask-SocketIO>=5.3.0  # For real-time WebSocket updates
eventlet>=0.33.0  # WSGI server for SocketIO
python-socketio>=5.8.0  # Socket.IO support
watchdog>=3.0.0  # For file system monitoring

# Authentication and security
Flask-JWT-Extended>=4.5.0  # JWT token authentication
bcrypt>=4.0.0  # Password hashing
itsdangerous>=2.1.0  # Secure token generation